+++ pkg/kube/container/containers.go
@@ -177,9 +177,16 @@ func WithPorts(ports []corev1.ContainerPort) Modification {
 	}
 }
 
-// WithSecurityContext sets teh container's SecurityContext
-func WithSecurityContext(context *corev1.SecurityContext) Modification {
+// WithSecurityContext sets the container's SecurityContext
+func WithSecurityContext(context corev1.SecurityContext) Modification {
 	return func(container *corev1.Container) {
-		container.SecurityContext = context
+		container.SecurityContext = &context
 	}
+}
+
+// DefaultSecurityContext returns the default container security context with:
+// - readOnlyRootFilesystem set to true
+func DefaultSecurityContext() corev1.SecurityContext {
+	readOnlyRootFilesystem := true
+	return corev1.SecurityContext{ReadOnlyRootFilesystem: &readOnlyRootFilesystem}
 }